Join Logman and Logman::Logger
