+++ lib/logman/version.rb
@@ -1,3 +1,3 @@
-module Logman
+class Logman
   VERSION = "0.1.0".freeze
 end
